Draft of a potential success page
Changed some button settings in the default.html layout. The only other file using this was exit-what-1.md which we determined we will remove later.

Had some questions about the back button here...should we redirect users back to the questions or at the start of the screener? Should we have the back button at all on this page??